+++ lib/pages/otherteampage.dart
@@ -76,7 +76,7 @@ class _OtherTeamPageState extends State<OtherTeamPage> {
 
         for (var item in data) {
           String name = item['name'];
-          int fundraiserBox = item['fundraiserBox'];
+          double fundraiserBox = item['fundraiserBox'];
           String? companyName;
 
           if (item['company'] != null) {
